Moved the spec tests
